backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-google-genai
yfinance
chromadb
beautifulsoup4
pandas
pandas-ta
feedparser